requirements.txt
kafka-python==1.4.4
prometheus_client==0.4.2
sqlalchemy==1.2.14
cx-Oracle==7.0.0
